Update for api keys

+++ .scripts/setup_generate_api_keys.sh
@@ -4,11 +4,21 @@ IFS=$'\n\t'
 
 setup_generate_api_keys()
 {
-    info "Generating API Keys"
-    echo "# API KEYS" > /opt/openflixr/api.keys
+    if [[ ! -f "/opt/openflixr/api.keys" ]]; then
+        touch "/opt/openflixr/api.keys"
+        echo "# API KEYS" > /opt/openflixr/api.keys
+    fi
+
+    info "Generating/Retrieving API Keys"
     for service in "${SERVICES[@]}"; do
         info "-- $service"
-        API_KEYS[$service]=$(uuidgen | tr -d - | tr -d '' | tr '[:upper:]' '[:lower:]')
-        echo "$service: ${API_KEYS[$service]}" >> /opt/openflixr/api.keys
+        if grep -q "^${service}" "/opt/openflixr/api.keys"; then
+            info "   Found API Key for ${service}. Retrieving..."
+            API_KEYS[$service]=$(grep "^${service}" "/opt/openflixr/api.keys" | cut -d " " -f 2)
+        else
+            API_KEYS[$service]=$(uuidgen | tr -d - | tr -d '' | tr '[:upper:]' '[:lower:]')
+            echo "${service}: ${API_KEYS[$service]}" >> /opt/openflixr/api.keys
+            info "   API Key created for ${service}"
+        fi
     done
 }
